backend/requirements.txt
psycopg2-binary==2.9.9
Faker==20.1.0
Flask
google-generativeai
python-dotenv
flask-cors

